update: renamed main navigation component
